Projeto 05.11.2024

+++ src/main/java/com/fatec/demo/controllers/ContaPagarRepository.java
@@ -3,5 +3,5 @@ package com.fatec.demo.controllers;
 import com.fatec.demo.entities.*;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContaPagarRepository extends JpaRepository<ContaPagar, Integer> {
+public interface ContaPagarRepository extends JpaRepository<ContaPagar, Long> {
 }
