+++ src/main/java/com/fatec/demo/entities/Fornecedor.java
@@ -12,7 +12,7 @@ import jakarta.persistence.Table;
 public class Fornecedor {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(nullable = false, length = 60)
     private String nome;
@@ -21,17 +21,17 @@ public class Fornecedor {
     public Fornecedor() {}
 
     // Construtor com parâmetros
-    public Fornecedor(int id, String nome) {
+    public Fornecedor(Long id, String nome) {
         this.id = id;
         this.nome = nome;
     }
 
     // Getters e Setters
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
